Year end encumbrance forward test is now 100% green.
Fixed ojb mapping for prior year account service. No primary key was defined and an indicator field in the bo was a String instead of a boolean as it should have been.
Rewrote rules for year end encumbrance closing process including the cobol code as comments for easy debugging.
Added a new constructor to OriginEntry to replace initializeEntry in (i think) the year end service.

Resolves KULGL-172 & KULGL-175.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@1445 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7