Changed fields in the form to use DataDictionary for the headers. I needed to ad d attributes and attribute references to the DataDictionary descriptor. I also added an attribute for the typeCode so that DataDictionary would populate the radio buttons.
Changes rules to use new import static from JDK 5.0.

Modified the forms to use David's refactoring for the Reversal Date

Had to create a special document for AuxiliaryVoucher when it is required to cre
ate a DistributionOfIncomeAndExpense. Since it bypasses the
user and any human human interface, the Document is required to route directly t
o FINAL.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@1868 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7