KULEDOCS-1567 KULEDOCS-342 -
Created StackRuntimeRule that is used for special case of comparing multiple things as a rule like objectType, subObjTyp, and objLevel at the same time. It is likely that when the rules framework is improved, that this will become unnecessary. It is just a temporary thing.

Made KualiConfigurationService more flexible to allow looking up a BusinessRule used to create a KualIParameterRule.

I fixed KULEDOCS-1566 by when I found the currentFiscalYear was being restricted as well as previous fiscal years.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@3349 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7