KULEDOCS-1647:
Only the description is required when saving a CCR doc.

Fixed minor error where the error message stating that at least one CCR must be added appeared in the other error section at the bottom of the form.  It has now been placed in the CCR tab as a section error.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@4153 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7