KULLAB-134 KULLAB-135 KULLAB-70 KULLAB-37
Changed an EventBase to an Event because it was not abstract by convention and no classes
extended it.

Since accounting lines are guaranteed to have good employee ids, the overrides in the
SalaryExpenseTransferAction are no longer needed.

Removing LaborDocument in favor of conventional LaborLedgerPostingDocumentBase

Made BenefitExpenseTransfer and SalaryExpenseTransfer LaborLedgerPostingDocuments

Added processCustomSaveDocumentBusinessRules to LaborExpenseTransferDocumentRuleBase to cause employee id to be validated on save.

Made it so that SalaryExpenseTransferDocumentAction performBalanceInquiry triggers a save to ensure that the employee id is properly
set before generating the report.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@7440 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7