KULLAB-134 KULLAB-135 KULLAB-70
The last checkin probably broke the build. Reverting back some code to fix it. Originally I was trying to avoid duplicating code, but got too deep and ended up duplicating code anyway. Without anyway around it, I'm checking in a fix.

Also, some areas were using AccountingDocument when it didn't seem applicable. For example, though AccountingDocuments have AccountingLines, they do not have LaborLedgerPendingEntries, so it doesn't make sense to generateLaborLedgerPendingEntries on an AccountingDocument. I changed this to LaborLedgerPostingDocument instead since LaborLedgerPostingDocument inherits from AccountingDocument. It made it easier to this way when rules were finally executed and needed to get to the pending entries. Normally, it would have had to cast, but now it doesn't.

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@7472 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7