KULLAB-291 KULLAB-292
I added a new getTypedService() method to SpringServiceLocator. This will make it so that we don't have to cast when using getService() anymore.

I changed references to AccountingLineBase in the AccountingLineDao/AccountingLineDaoOjb to AccountingLine to be more consistent with the "programming-to-interface" methodology.

I created the AccountingDocumentHelper in order to get functionality out of AccountingDocument for other documents with AccountingDocument characterstics (like AccountingLines,) but cannot inherit from AccountingDocument because of the Hierarchy. This may become redundant with my upcoming proposal on an AccountingDocument change.

As a result of the above, I had to add custom events and rules into Labor. Those may also be removed later.

I added a new service that injects ExpenseTransferAccountingLineDaoOjb to get around KULLAB-839

git-svn-id: https://svn.kuali.org/repos/kfs/trunk@8207 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7